requirements.txt
httpx==0.26.0
python-telegram-bot==20.8
pytz==2024.1
ta==0.11.0
pandas==2.2.2
numpy==1.26.4
schedule==1.2.1
aiofiles==23.2.1
types-pytz
types-aiofiles
